Source folder reorganization and initial main class/configuration files.
Note that by default, the server starts listening with SSL using a self-
signed, and (because both it and its key are in the repository) very insecure,
certificate. This is solely to ensure that we can successfully use SSL, is not
meant for deployment, and may be removed in a later commit.