defs added based on SS's domain knowledge
I've departed radically from Joe's definition of "ballot". Joe's def was: "The legally recognized means by which a voter can express his or her choices in one or more contests in an election." Mine is "a list of contests and, for each contest in the list, a list of choices, in a form that allows a person or computer to record choices, read recorded choices, or both" 

What I like about my definition is that it makes "Paper Ballot" and other items easier to define. I also worry that Joe's definition pulls a lot of other things in, such as the voting machine itself.