Add smoketest of cvr, manifest uploads to server
At this point this is very messy, with some very very long lines of static data in auto-generated code,
no error handling in the main.py upload code, etc.

But it runs a smoketest, testing several of the endpoints in corla-server-test.html,
and provides a basis for comment and refinement.