Add Dockerfile: config aid for CentOS 7 / RHEL 7
The Dockerfile builds a package with all the CentOS/RHEL packages needed
for the server.
It serves as a test of doing parts of that install, as well as a form of
documentation of the exact packages involved.

The associated docker/README.md file documents some of the additional
manual steps used to build a working docker container.

And of course, it is a good part of the way towards actually supporting
an automated Docker build, for Q/A, CI, and deployment purposes.

TODO:
 Determine the best places to put these files.
 Automate the rest of the setup process.
 Follow other docker best practices in terms of naming, tags, MAINTAINERs etc.
 Tighten down security, by running as a non-root user and adding other restrictions.