Fix audit CVR upload.
The object `mark.choices` maps choice names to truth values. We should
include a choice (key name) in the contest info array iff its value in
the object is `true`. We failed to filter first, and so we included
any key that existed in the object. This resulted in correct behavior
if that choice's associated checkbox had never been checked, because
the key would not appear in the object. If the choice had been checked
and then unchecked, the behavior would be wrong, because the existence
of a key for the choice (though mapped to `false`) would still be
enough to include the choice in the `choices` array of the contest
info.