Fix computation of round completion ratio.
Closes #728.

If a county had participated in a round, but had completed the round,
we were excluding them from the count of total participants. Now, we
obtain this value by checking if the county had participated in the
round at any point, and then check each county's current round to
determine if they are done with the round or not.