Pass auditability predicate function to select contest form.
Also ensure that we don't consider a contest auditable if it has type
`HAND_COUNT`. This should never occur in the select contest form, but
this way, if there is a bug and it does, we won't let the user try to
select it, only to fail confusingly upon submission.