Set county statuses in DOS dashboard refresh, don't merge.
Closes #790.

We sometimes failed to detect that a county had completed a round.
This is because we derive that fact from (1) the presence of the
current round in the county's round history, (2) the absence of that
round from its `currentRound`. However, in some cases, the round would
be over, but the current round would still be nonempty.

This is because in the DOS dashboard refresh JSON, a county's new
`current_round` from the server would be `null`, so deeply merging it
overb `obj` gives us `obj` instead of clearing it.

Things would seem correct if we refreshed, because in that case, we'd
start with the `null`, and future refreshes would either merge `null`
over `null` (yielding a correct `null` value) or `obj` over `null`
(yielding a correct `obj` value). However, the new `countyStatus` we
receive is always complete, and doesn't need to be merged, so we just
set it instead.