Make component selectors expect their dashboard's state.
We used to have a single state type, and each dashboard had its own
nullable key inside of that type. Now, we have distinct types of app
state in a union. We still need to statically ensure a connected
component will never be rendered with the wrong state variant, but
this is a start in that direction.