Lowercase username parameter to login2F
As we observed in CO, users attempting to log in with a non-lowercase
username were stymied on the second factor log-in form. When the user
enters their username as lowercase, both first- and second- factor forms
work as expected. It stands to reason that we need to lowercase the
username on both forms, since that would provide the same result as if
the user had initially entered their username in lowercase.