Fix a bug where standardized contests would not display on first run
The current architecture of the audit definition workflow is highly
asynchronous, to the point that we render the standardized contests page
before the file upload is even complete. If no data is present in the
database, it is possible to have renders that change standardized
contest props more than two times but without new data, thus fooling the
brittle flag-checks we were doing.

When I wrote this, I did not realize that the standardized contest
upload was a required step. We can now leverage that knowledge to wait
until there is actually data before making a decision as to whether we
want to skip the contest standardization step or not, rather than
relying on props changing.