[wip] A sketch of making a larger comparison audit
The problem we're trying to solve is discrepancies being kept separate
in little county buckets when auditing a multi-county (from 2..64)
contest. Previously, we had a bunch of CountyContestComparisonAudit
objects that weren't combined, but we need to consider discreps over the
set of counties participating in some contest audit.

There is a lot yet to be done:
- figure out how to look at auditPrefixLength over many dashboards
- get the UI to show a ContestResult view
- lots more that lives in FIXME and TODO

You should ignore whitespace when you view the diffs!

This is a mashup of some previous work:
- [wip] sketch of removing the CountyContest part of ComparisonAudit
- [wip] refactoring attempt and notes
- [wip] when this compiles, we should know if this approach will work