Revise v1.1 rla_export reports and report names
Combine the all_contest_dynamic and all_contest_static reports into a
single report with an easier name: contest.  That has the risk_limit
column and all the others, so there is a single comprehensive way to
view the data that fed into the audit and that comes out of it -
exactly what observers will be most interested in.

Eliminate the m_selected_contest_dynamic and m_selected_contest_static
reports which were just subsets of the ones just mentioned.

Rename the all_contest_audit_details_by_cvr to contest_comparisons,
and introduce a new report without the choice_per_voting_computer
column, calling it contest_details.

Rename the rest of the m_* reports, simply removing the unnecessary
and confusing prefix which was there for historical reasons.