Don't try to handle ties in a County context
Now that we have contests spanning counties, we need to wait until all
possible CVR exports have been uploaded before we can count the results
and look for any ties.

By doing it at the time of CVR import, a contest that might have tied in
a county - but not overall - gets placed in the
DoSDashboards.contestsToAudit() when it has no business being there yet.