Be a little stricter when ending audits by county
If County A and County B were both participating in some multi-county
`Contest One` and County B had a second single-county `Contest Two`, the
simple act of running out of ballots to audit in County B (low voter
turnout?) would try to end all of County B's audits. That's fine for
`Contest Two`, but if we end `Contest One`, then we'll never be able to
achieve a risk limit and complete as hoped.