A script to reproduce a failed second round
At this point in commit-order, there should an exception about state
transitions. This scenario gives no work to two counties in the first
round. When a round is started, it needs to be ended because there's a
check of `my_current_round_index` in `CountyDashboard/startRound`. (See
also `StartAuditRound/startRound:393`)