Update audit status when counting previously auditied samples
An audit's status is updated when an ACVR is submitted. This is great,
but if we start a new round and draw previously-audited samples, we
won't know if we've seen enough samples to reach our risk limit. In an edge
cases where these high-multiplicity samples are at the the tail end of a
county's ballot sequence, we'll end the round without having updated the
status. This could lead to infinite rounds of nothing to do.